requirements/base.txt
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
Flask==3.1.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pika==1.3.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
redis==5.2.1
requests==2.32.3
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
redlock-py==1.0.8
Flask-Pydantic==0.13.1
msgpack==1.1.0
